Break out classes into a separate module and further cleanup
